fix: proxy module
