+++ service/vspo-schedule/proxy/pkg/processor/index.ts
@@ -1,0 +1,2 @@
+export * from './event';
+export * from './video';
