+++ service/vspo-schedule/proxy/schema.ts
@@ -5,13 +5,13 @@ export const VideoSchema = z.object({
     id: z.string(),
     title: z.string(),
     description: z.string(),
-    channelId: z.string(),
-    channelTitle: z.string(),
-    thumbnailUrl: z.string(),
-    scheduledStartTime: z.string(),
+    channelId: z.string().optional(),
+    channelTitle: z.string().optional(),
+    thumbnailUrl: z.string().optional(),
+    scheduledStartTime: z.string().optional(),
     actualEndTime: z.string().optional(),
-    iconUrl: z.string(),
-    platform: z.enum(["youtube", "twitch", "twitcasting", "nicovideo"]),
+    iconUrl: z.string().optional(),
+    platform: z.enum(["youtube", "twitch", "twitcasting", "nicovideo"]).optional(),
     link: z.string().optional(),
     viewCount: z.number().optional(),
     likeCount: z.number().optional(),
@@ -22,3 +22,13 @@ export const VideoSchema = z.object({
     isTemp: z.boolean().optional(),
     tempUrl: z.string().optional(),
 });
+
+export const EventSchema = z.object({
+    title: z.string(),
+    contentSummary: z.string(),
+    webPageLinks: z.array(z.string()),
+    tweetLinks: z.array(z.string()),
+    startedAt: z.string(),
+    isNotLink: z.boolean(),
+    newsId: z.string(),
+})
